Captioning/requirements.txt
torch>=1.9
transformers
tqdm
opencv-python
Pillow
einops
